### 2016-06-09 #26 @whedon
Submission: Armadillo: a template-based C++ library for linear algebra
**Submitting author:** @conradsnicta (<a href="http://orcid.org/0000-0002-0049-4501">Conrad Sanderson</a>)
**Repository:** <a href="https://github.com/conradsnicta/armadillo/" target ="_blank">https://github.com/conradsnicta/armadillo/</a>
**Version:** v7.200.1b
**Archive:** <a href="http://dx.doi.org/10.5281/zenodo.55251" target ="_blank">10.5281/zenodo.55251</a>

- [x] **Version:** Does the release version given match the GitHub release (v7.200.1b)?
- [x] **Archive:** Does the <a target="_blank" href="http://dx.doi.org/10.5281/zenodo.55251">software archive</a> resolve?

- [x] **Functionality documentation:** Is the functionality of the software documented to a satisfactory level (e.g. API method documentation)?
