### 2016-06-09 #26 @wrathematics
Done reviewing. I have only a few minor comments.

The DOI for the software repository (incorrectly) lists the license as MPL 1.1, rather than MPL 2.0. This may seem minor, but it's worth nothing that MPL 1.1 is not an OSI approved license, though MPL 2.0 is. I actually had a mild moment of panic when I saw that, because my understanding is MPL 1.1 isn't GPL compatible, which would affect a large number of current users. But both the `LICENSE.txt` and `README.txt` files confirm it is being distributed under MPL 2.0.

The only other possible quibble is the performance checkbox (which I did check). No specific performance claims are made, but several vague claims are made to the general effect that it achieves high performance. I have used Armadillo in the past and that is my general impression, although I have not tested this extensively and no benchmarks are provided. I personally would like to see at least some simple benchmarks, for example comparing the overhead of using Armadillo to compute an SVD over calling LAPACK directly. Though that may be beyond the scope of this review, and probably shouldn't be a strict requirement.

I recommend publication, ideally after fixing the license statement in the DOI.
