### 2016-06-09 #26 @arfon
Thanks for the rapid turnaround @wrathematics!

> I recommend publication, ideally after fixing the license statement in the DOI.

@conradsnicta - can you reply on this thread when you've had a chance to address this?
