### 2016-06-10 #26 @conradsnicta
@arfon - I've fixed the license metadata on the Zenodo archive. It now states "License: Other (Open)". Zenodo currently doesn't have an option for MPL v2.0, so this is the closest choice.

Also, I've incorporated a few minor fixes to the references (paper.bib). Could you pull them in and regenerate the paper PDF?
