### 2016-06-10 #26 @arfon
Updated paper: [10.21105.joss.00026.pdf](https://github.com/openjournals/joss-reviews/files/309093/10.21105.joss.00026.pdf)
