### 2016-06-10 #26 @arfon
@conradsnicta - I think this is good to accept now. Are there any other changes you're planning on making?

/ cc @wrathematics
