### 2016-06-10 #26 @arfon
> However, there seems to be an inconsistency in the referencing style in the generated paper (page 2). Specifically, the ordering of the names doesn't appear right: the first author in each reference is listed as Lastname, Firstname, and then following authors are listed as Firstname, Lastname. Is this deliberate?

Yeah, I'd spotted that too. I'm just using the default bibliographic style from LaTeX/Pandoc.
