### 2016-06-10 #26 @conradsnicta
Maybe try this in the preamble of paper.tex:
    `\usepackage{natbib}`
    `\setcitestyle{authoryear,open={(},close={)}}`
then use `\citep{refname]` inside the body for citations.
Finally, just before `\end{document}`:
      `\bibliographystyle{plainnat}`
      `\bibliography{paper}`
