### 2016-06-10 #26 @arfon
Thanks @conradsnicta, I tried adding most of this to the [LaTeX template](https://github.com/openjournals/whedon/blob/master/resources/latex.template) but it didn't seem to change the PDF output.

> then use `\citep{refname}` inside the body for citations.

I don't think we can do this as we go from the Markdown to PDF (via LaTeX) in a single step [here](https://github.com/openjournals/whedon/blob/master/lib/whedon/processor.rb#L152-L171) with the citations being handled by pandoc.

I'm afraid we've reached the limit of my pandoc-fu :-\. I'd propose that we accept this as-is and file an issue on https://github.com/openjournals/whedon about this. Does that sound OK?
