### 2016-06-10 #26 @conradsnicta
@arfon - sounds good.  Please go ahead.
