### 2016-06-10 #26 @conradsnicta
@arfon - issue filed: https://github.com/openjournals/whedon/issues/2
