### 2016-06-10 #26 @arfon
@wrathematics many thanks again for the rapid review here!

@conradsnicta - the DOI for this paper is http://dx.doi.org/10.21105/joss.00026 🎉 🚀 :boom:. The URL doesn't resolve just yet as the Crossref submission queue is a little backed up.
